chore: deprecate AbstractConfigurationBuilder#withParentClasspath(Class<?> klass, FileFilter file)
With JDK9, the application classloader does not inherit from
URLClassLoader. So, it becomes impossible to create a new
URLClassLoader from it, by filtering classpath entries. So, this
method is now deprecated and will be removed in a next release.

For any issue, please submit an issue on the bug tracker to discuss it.